MA-18798-3 power: Fix Unintentional integer overflow.
this patch is to fix [Coverity CID :11903894] "Unintentional integer
overflow(OVERFLOW_BEFORE_WIDEN)"

Potentially overflowing expression s_previous_duration * 1000 with type
int (32 bits, signed) is evaluated using 32-bit arithmetic, and then used
in a context that expects an expression of type long long (64 bits, signed).

to avoid overflow, cast either s_previous_duration or 1000 to type long long.

Reviewed-on: http://androidsource.ap.freescale.net/project/15269
(cherry picked from commit a338c740195dc641e461cd7bdced52c70b88e741)
Reviewed-on: http://androidsource.ap.freescale.net/project/15318

+++ power/Power.cpp
@@ -177,7 +177,7 @@ ndk::ScopedAStatus Power::setBoost(Boost type, int32_t durationMs) {
 
             long long elapsed_time = calc_timespan_us(s_previous_boost_timespec, cur_boost_timespec);
             // don't hint if previous hint's duration covers this hint's duration
-            if ((s_previous_duration * 1000) > (elapsed_time + duration * 1000)) {
+            if (((long long)s_previous_duration * 1000) > (elapsed_time + duration * 1000)) {
                 break;
             }
